+++ lib/applitools/appium/driver.rb
@@ -2,10 +2,11 @@
 
 module Applitools::Appium
   class Driver < Applitools::Selenium::Driver
-    attr_accessor :appium_driver
     def initialize(eyes, options)
-      self.appium_driver = options.delete(:appium_driver)
       super(eyes, options)
+    end
+    module AppiumLib
+      extend self
     end
   end
 end