+++ lib/applitools/selenium/eyes_target_locator.rb
@@ -210,7 +210,7 @@ module Applitools::Selenium
         when String
           frame(name_or_id: frame_name_or_id)
         when Applitools::Selenium::Element
-          frame(frame_element: frame_name_or_id.reference)
+          frame(frame_element: frame_name_or_id)
         else
           Applitools::ArgumentGuard.raise_argument_error Applitools::EyesNoSuchFrame.new frame_name_or_id
         end